+++ utils/logging.py
@@ -1,0 +1,29 @@
+import logging
+from opentelemetry._logs import set_logger_provider
+from opentelemetry.sdk._logs import LoggerProvider, LoggingHandler
+from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
+from azure.monitor.opentelemetry.exporter import AzureMonitorLogExporter
+import os
+
+def setup_logging():
+    # Set up logs export to Azure Application Insights
+    logger_provider = LoggerProvider()
+    set_logger_provider(logger_provider)
+    exporter = AzureMonitorLogExporter(
+        connection_string=os.environ["APPLICATIONINSIGHTS_CONNECTION_STRING"]
+    )
+    logger_provider.add_log_record_processor(BatchLogRecordProcessor(exporter))
+
+    # Attach LoggingHandler to root logger
+    handler = LoggingHandler()
+    logging.getLogger().addHandler(handler)
+    logging.getLogger().setLevel(logging.NOTSET)
+    logger = logging.getLogger(__name__)
+
+    # Silence noisy loggers
+    logging.getLogger("requests").setLevel(logging.WARNING)
+    logging.getLogger("urllib3").setLevel(logging.WARNING)
+    logging.getLogger("azure").setLevel(logging.WARNING)
+    logging.getLogger("requests_oauthlib").setLevel(logging.WARNING)
+    logging.getLogger("asyncio").setLevel(logging.WARNING)
+    logging.getLogger("opentelemetry").setLevel(logging.ERROR)
